requirements.txt
asgiref==3.2.10
Django==3.0.8
django-cors-headers==3.4.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
gcloud==0.18.3
googleapis-common-protos==1.52.0
httplib2==0.18.1
mysqlclient==2.0.1
oauth2client==4.1.3
protobuf==4.0.0rc2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
rsa==4.6
six==1.15.0
sqlparse==0.3.1
